interface matching

+++ lv_cpp/widgets/LvTable.cpp
@@ -66,7 +66,7 @@ bool LvTable::hasCellCtrl(uint16_t row, uint16_t col, lv_table_cell_ctrl_t ctrl)
 	return lv_table_has_cell_ctrl(cObj.get(),row,col,ctrl);
 	
 }
-LvTable& LvTable::getSelectedCell(uint16_t *row, uint16_t *col){
+LvTable& LvTable::getSelectedCell(uint32_t *row, uint32_t *col){
 	lv_table_get_selected_cell(cObj.get(),row,col);
 	return *this;
 }
